Add pluggable password validation.
Password validation for registration, change, and reset has only ever
checked for length - and the length wasn't configurable.

This add an easy way for applications to provide their own password validator as
well as provide a default password validator that checks:

1) length
2) complexity using zxcvbn
3) breaches using https://haveibeenpwned.com/API/v3

These 3 validations are also made available as part of the API so that applications
that need their own validator can still make use (if desired) of the underlying validators.

New error messages and configurations were added and documented.

The default minimum password length has been changed from 6 to 8 (NIST).

The one downside to using zxcvbn is that it is no longer maintained (the original JS version)
and the python version currently isn't very string-localizable friendly.

closes: #195
closes: #99